more soak tests
